Pull the body of influx_stress main out
This commit abstracts out the body of `main()` and moves it into a separate
function `Runner` in `runner/runner.go`. Additonally two new types,
`Timer` and `Config` were introduced.

`Runner` takes in a `*Config` and returns the total number of points
written `totalPoints`, an slice of response times `responseTimes`, and a
timer with the starting and ending times of the test `*Timer`.
